move equipment admin form to harvest/admin.py
