+++ saskatoon/harvest/admin.py
@@ -5,6 +5,9 @@ from django.db.models import Value
 from django.db.models.functions import Replace
 from django.urls import reverse
 from django.utils.html import mark_safe
+from django.utils.translation import gettext_lazy as _
+from django import forms
+from dal import autocomplete
 from harvest.models import (Property, Harvest, RequestForParticipation, TreeType,
                             Equipment, EquipmentType, HarvestYield, Comment,
                             PropertyImage, HarvestImage)
@@ -79,6 +82,31 @@ class RequestForParticipationAdmin(admin.ModelAdmin):
 @admin.register(Equipment)
 class EquipmentAdmin(admin.ModelAdmin):
     form = EquipmentAdminForm
+
+
+class EquipmentAdminForm(forms.ModelForm):
+    def clean(self):
+        cleaned_data = super(EquipmentAdminForm, self).clean()
+        bool1 = bool(self.cleaned_data['property'])
+        bool2 = bool(self.cleaned_data['owner'])
+        if not (bool1 != bool2):
+            raise forms.ValidationError(
+                _('Fill in one of the two fields: property or owner.')
+            )
+        return cleaned_data
+
+    class Meta:
+        model = Equipment
+        widgets = {
+            'property': autocomplete.ModelSelect2(
+                'property-autocomplete'
+            ),
+            'owner': autocomplete.ModelSelect2(
+                'actor-autocomplete'
+            ),
+        }
+
+        fields = '__all__'
 
 
 class PropertyImageInline(admin.TabularInline):
