+++ saskatoon/harvest/forms.py
@@ -608,30 +608,6 @@ class HarvestYieldForm(forms.ModelForm):
             ),
         }
 
-class EquipmentAdminForm(forms.ModelForm):
-    def clean(self):
-        cleaned_data = super(EquipmentAdminForm, self).clean()
-        bool1 = bool(self.cleaned_data['property'])
-        bool2 = bool(self.cleaned_data['owner'])
-        if not (bool1 != bool2):
-            raise forms.ValidationError(
-                _('Fill in one of the two fields: property or owner.')
-            )
-        return cleaned_data
-
-    class Meta:
-        model = Equipment
-        widgets = {
-            'property': autocomplete.ModelSelect2(
-                'property-autocomplete'
-            ),
-            'owner': autocomplete.ModelSelect2(
-                'actor-autocomplete'
-            ),
-        }
-
-        fields = '__all__'
-
 
 class EquipmentForm(forms.ModelForm):
     class Meta:
